pyproject.toml
[project]
name = "hwp-ai-autofill"
version = "0.1.0"
description = "웹 채팅 인터페이스로 정부지원사업 계획서 HWP 템플릿을 AI로 자동 채워넣는 시스템"
requires-python = ">=3.11"
readme = "CLAUDE.md"
license = { text = "MIT" }

dependencies = [
    # HWP 처리
    "pyhwpx>=0.5",
    "pywin32>=306",
    "python-hwpx>=0.1",
    "lxml>=5.0",

    # 웹 서버
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "websockets>=13.0",
    "python-multipart>=0.0.12",

    # LLM 연동
    "anthropic>=0.40",
    "openai>=1.50",
    "litellm>=1.50",
    "httpx>=0.27",

    # 데이터/유틸리티
    "structlog>=24.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "typer>=0.12",
    "pandas>=2.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",

    # RAG
    "chromadb>=0.5",

    # MCP (Claude Desktop / Cursor 연동)
    "mcp[cli]>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
    "pytest-asyncio>=0.24",
]

[project.scripts]
hwp-ai = "src.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
